requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beartype==0.21.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
bs4==0.0.2
by==0.0.7
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cramjam==2.11.0
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
ddgs==9.5.5
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docx==0.2.4
docx2txt==0.9
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fake-useragent==2.2.0
fastai==2.8.4
fastcore==1.8.8
fastdownload==0.0.7
fastjsonschema==2.21.1
fastparquet==2024.11.0
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.19.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
icrawler==0.6.10
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.44
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.0
lxml_html_clean==0.4.3
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
openai==2.3.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pdfminer==20191125
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
platformdirs==4.3.8
playwright==1.55.0
plum-dispatch==2.5.7
polars==1.33.0
preshed==3.0.10
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
pypdf==6.1.3
PyPDF2==3.0.1
pypdfium2==5.0.0
pyppeteer==2.0.0
pyquery==2.0.1
pyresparser==1.0.6
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-html==0.10.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tdqm==0.0.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
torch==2.8.0
torchaudio==2.8.0+cpu
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
tqdm-pandas==1.7.0
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.17.4
types-python-dateutil==2.9.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.196
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
w3lib==2.3.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
